split the guess logic in its separate file
The guessing logic will probably never be used again, so it's better to split it into its own separate file. The file was also too long (700 lines..) So now we have wadecrypt that checks every single byte of the input and waguess that just tries to decrypt stuff ignoring everything.